More Race Structure & Fixed Double Collisions
It was the mouse collider. Unchecked a property and it worked.

Also did a bunch more work to form the race structure. I'm still not sure what the role of the titular "race director" is, found myself trying to automate certain parts of the flag colors and all...should I?